[reconfig] Pass epoch id from ConsensusHandler
